requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0


